+++ com/naukaJavy/Bilbo/Treasure.java
@@ -1,13 +1,13 @@
 package com.naukaJavy.Bilbo;
 
 public class Treasure {
-    private int size;
+    private float size;
 
-    public int getSize() {
+    public float getSize() {
         return size;
     }
 
-    public Treasure(int size) {
+    public Treasure(float size) {
         this.size = size;
     }
 }
